--- a/src/init.cpp
+++ b/src/init.cpp
@@ -1484,6 +1484,24 @@ bool AppInitMain(NodeContext& node, interfaces::BlockAndHeaderTipInfo* tip_info)
         node.chainman = std::make_unique<ChainstateManager>(chainman_opts, blockman_opts);
         ChainstateManager& chainman = *node.chainman;
 
+        // This is defined and set here instead of inline in validation.h to avoid a hard
+        // dependency between validation and index/base, since the latter is not in
+        // libbitcoinkernel.
+        chainman.restart_indexers = [](const ChainstateManager& chainman) {
+            // Drain the validation interface queue to ensure that the old indexers
+            // don't have any pending work.
+            SyncWithValidationInterfaceQueue();
+
+            for (auto* indexer : chainman.indexers) {
+                indexer->Interrupt();
+                indexer->Stop();
+                if (!indexer->Start()) {
+                    // TODO I don't think we should halt here, but maybe do something more drastic?
+                    LogPrintf("[snapshot] WARNING failed to restart indexer %s on snapshot chain\n", indexer->GetName());
+                }
+            }
+        };
+
         node::ChainstateLoadOptions options;
         options.mempool = Assert(node.mempool.get());
         options.reindex = node::fReindex;