--- a/src/interfaces/chain.h
+++ b/src/interfaces/chain.h
@@ -27,6 +27,7 @@ class Coin;
 class uint256;
 enum class MemPoolRemovalReason;
 enum class RBFTransactionState;
+enum class ChainstateRole;
 struct bilingual_str;
 struct CBlockLocator;
 struct FeeCalculation;
@@ -271,9 +272,10 @@ public:
         virtual void transactionAddedToMempool(const CTransactionRef& tx) {}
         virtual void transactionRemovedFromMempool(const CTransactionRef& tx, MemPoolRemovalReason reason) {}
         virtual void blockConnected(const BlockInfo& block) {}
+        virtual void blockConnected(const ChainstateRole role, const BlockInfo& block) {}
         virtual void blockDisconnected(const BlockInfo& block) {}
-        virtual void updatedBlockTip() {}
-        virtual void chainStateFlushed(const CBlockLocator& locator) {}
+        virtual void updatedBlockTip(const ChainstateRole role) {}
+        virtual void chainStateFlushed(const ChainstateRole role, const CBlockLocator& locator) {}
     };
 
     //! Register handler for notifications.