--- a/src/net_processing.cpp
+++ b/src/net_processing.cpp
@@ -1876,8 +1876,14 @@ void PeerManagerImpl::BlockConnected(
     const std::shared_ptr<const CBlock>& pblock,
     const CBlockIndex* pindex)
 {
-    m_orphanage.EraseForBlock(*pblock);
+    // Update this for all chainstate roles so that we don't mistakenly see peers
+    // helping us do background IBD as having a stale tip.
     m_last_tip_update = GetTime<std::chrono::seconds>();
+
+    if (role == ChainstateRole::BACKGROUND) {
+        return;
+    }
+    m_orphanage.EraseForBlock(*pblock);
 
     {
         LOCK(m_recent_confirmed_transactions_mutex);
@@ -1979,6 +1985,9 @@ void PeerManagerImpl::NewPoWValidBlock(
  */
 void PeerManagerImpl::UpdatedBlockTip(const ChainstateRole role, const CBlockIndex *pindexNew, const CBlockIndex *pindexFork, bool fInitialDownload)
 {
+    if (role == ChainstateRole::BACKGROUND) {
+        return;
+    }
     SetBestHeight(pindexNew->nHeight);
     SetServiceFlagsIBDCache(!fInitialDownload);