--- a/src/validation.cpp
+++ b/src/validation.cpp
@@ -5850,3 +5850,11 @@ bool ChainstateManager::ValidatedSnapshotCleanup()
     }
     return true;
 }
+
+
+bool ChainstateManager::IsAnyChainInIBD() const
+{
+    return
+        (m_snapshot_chainstate && m_snapshot_chainstate->IsInitialBlockDownload()) ||
+        (m_ibd_chainstate && m_ibd_chainstate->IsInitialBlockDownload());
+}