… doc/release-process.md
… src/bench/wallet_create_tx.cpp
… src/node/blockstorage.cpp
… src/node/blockstorage.h
… src/node/chainstate.cpp
… src/node/interfaces.cpp
… src/rpc/blockchain.cpp
… src/rpc/mining.cpp
… src/test/blockmanager_tests.cpp
… src/test/coinstatsindex_tests.cpp
… src/test/fuzz/rpc.cpp
… src/test/util/validation.cpp
… src/test/util/validation.h
… src/test/validation_block_tests.cpp
… src/test/validation_chainstatemanager_tests.cpp
… src/test/validationinterface_tests.cpp
… src/validation.h
… src/validationinterface.cpp
… src/validationinterface.h
… src/wallet/test/fuzz/notifications.cpp
… src/wallet/wallet.cpp
… src/wallet/wallet.h
… src/zmq/zmqnotificationinterface.cpp
… src/zmq/zmqnotificationinterface.h
… test/functional/feature_assumeutxo.py
… test/functional/test_runner.py
… test/lint/lint-shell.py