@gmaxwell 2019-03-15
> Presumably the snapshot would be taken relatively close to the current tip but far enough away to avoid meaningful reorgs, say 10,000 blocks deep.

To be clear, the snapshot can't be terribly recent for the user of it or it breaks the security model. Assume valid only has any security at all if there is time for the review and communication about review to happen, which means a human timescale of at least weeks.

Probably the assumption should be that snapshots are created pretty close to tip, just far enough to avoid wasting time in reorgs (even 6 blocks would be fine) but the users aren't using them until they are quite a bit older, likely using the second to most recent available one.

> We don't serve blocks or transactions to the network while we're operating with an unvalidated snapshot-based chain.

We shouldn't do that.  The worst that getting it wrong does is getting us disconnected from peers, which isn't particularly cosmic (and at least we might notice that something is going wrong).  The downside of it is that not forwarding transactions shoots our privacy in the head, since any transaction we're emitting would be one we created.

> When we attempt to load a wallet with a BestBlock locator lower than the base of a snapshot and the snapshot has not yet been validated, we refuse to load the wallet.

Sounds good.

>  compare it with the claimed hash in the snapshot metadata.

This is a zero security approach.  You can't just ask the user for a value and then accept that. The attack is to reorg the chain and then announce to everyone that there is a node bug and you need to run this command to continue, we specifically engineered out that possibility in assumevalid.  This isn't hypothetical, this is exactly what we've seen happen in ethereum w/ fastsync.

This is a fine setup for testing your PR however!    Ultimately we should get it to a state where a (FEC-split) copy of the state is loaded from the network and where it can be tested against a publicly reviewed constants configured in the software and/or blockchain.