@jamesob 2019-03-19
> Probably the assumption should be that snapshots are created pretty close to tip, just far enough to avoid wasting time in reorgs (even 6 blocks would be fine) but the users aren't using them until they are quite a bit older, likely using the second to most recent available one.

Yep - I figured that we'd update the assumeutxo hash in lockstep with assumevalid's.

> We shouldn't do that. The worst that getting it wrong does is getting us disconnected from peers, which isn't particularly cosmic (and at least we might notice that something is going wrong). The downside of it is that not forwarding transactions shoots our privacy in the head [...]

Good points, will fix that.

> This is a zero security approach. You can't just ask the user for a value and then accept that.

Agreed - the check I do there is just for sanity. Deciding on an initial hardcoded `assumeutxo` hash seems corequisite to allowing snapshots to be loadable through RPC (or any other means).