@jamesob 2019-03-30
Thanks for the suggestion, @ryanofsky. I spent the last few days reconstructing the changeset into sensible commits - hopefully it's easier to understand this way.

Most of the early commits are just shuffling stuff around (though all of it strictly necessary AFAICT). I've phrased the changes as much as possible as scripted-diffs and move-onlys. In replaying the changes, I found a few unnecessary diffs to omit.

If anyone has additional suggestions for how this could be made easier to review, I'm all ears, though I'm not holding my breath waiting until some Concept (N)ACKs roll in on #15605.