@Sjors 2019-04-19
I'm unable to compile this on macOS. See [error log](https://gist.github.com/Sjors/3937c8ebef0fcc8ea066072eacb59a60). It does compile on Ubuntu, despite quite a few warnings.

I tested making a dump for testnet and then importing it with a fresh datadir. But I think I'm using it wrong. First time I tried to load it I got. That was before it fetched any headers. I waited for it to fetch a few blocks and tried again, but same error:

```
bitcoind: txdb.cpp:104: virtual bool CCoinsViewDB::BatchWrite(CCoinsMap&, const uint256&): Assertion `!hashBlock.IsNull()' failed.
```

I don't think an RPC method is the right tool for loading a snapshot (it's fine for saving); probably better to just specify a location at launch.

Can you add an actual `assumeutxo` value (mainnet & testnet)? When combined with #13713 it should be easier to create and test the snapshot (I had some difficulty wrapping my head around the demo script). That also means `loadtxoutset` can refuse to process a snapshot for a lower block height than the most recent fully processed block (even if the header is a valid-fork).

To facilitate writing functional tests, you could allow a `-assumeutxo` param for regtest only.