@jamesob 2019-04-23
> I'm unable to compile this on macOS. See error log. It does compile on Ubuntu, despite quite a few warnings.

Thanks; could you post the warnings you're receiving on Ubuntu? I'm not getting any warnings during compilation when using clang 9.0.0 and gcc 7.3.0 on Ubuntu 18.04.2.

> I don't think an RPC method is the right tool for loading a snapshot (it's fine for saving); probably better to just specify a location at launch.

Loading snapshot only on startup would simplify things in the short run, but ultimately if we're going to share snapshots over the P2P network, we'll need to support loading them at runtime. Better to test that process out earlier IMO.

> I waited for it to fetch a few blocks and tried again, but same error...

Oops, looks like I introduced a bug during a recent rebase. Fixing now. Thanks for testing!

> Can you add an actual assumeutxo value (mainnet & testnet)?

Yep, I'll do this in the next day or so.

> To facilitate writing functional tests, you could allow a -assumeutxo param for regtest only.

Yep, this seems like a good idea. I'm also going to try to do as much unittesting as I can (as opposed to functional tests).