@jamesob 2019-04-25
I've added particular assumeutxo values and a more narrow `devtools` script. @Sjors hopefully this should now build without trouble or warnings on MacOS, though I haven't tested.