@jamesob 2019-04-26
When (and if) I get a few Concept ACKs on this, I'll start to break this into smaller PRs with accompanying tests. Let me know if there's anything that would make conceptual review easier.