@jamesob 2019-11-18
A few quick updates on this branch.

### New snapshots

I've added an allowed assumeutxo hash at height 600,000 (which you can check with `./contrib/devtools/utxo_snapshot.sh - ./src/bitcoin-cli -datadir=[...]`) and have uploaded snapshot files ([Google Drive](https://storage.googleapis.com/assumeutxo/utxo.600k.dat), [Google Cloud](https://drive.google.com/file/d/1sGNpAsdQPPZNpWt7Wwbi5u3bHZiIKmSb/view?usp=sharing)) for testers' convenience. We obviously don't necessarily need to preserve this value for merge, but I figure it's useful for testing and not unrealistic.

You can test this branch easily by running a script like
```sh
#!/bin/bash

set -e
if [ ! -f ./src/bitcoind ]; then
  echo "Must be in built bitcoin directory."
  exit 1
fi
SNAPSHOT_LOC=/tmp/utxo.600k.dat
DATADIR=/tmp/bitcoin-assumeutxo-test
mkdir -p $DATADIR

[ -f ${SNAPSHOT_LOC} ] || curl -L https://storage.googleapis.com/assumeutxo/utxo.600k.dat > $SNAPSHOT_LOC
/usr/bin/time -v ./src/bitcoind -datadir=$DATADIR -dbcache=4000 -stopatheight=604401 &
sleep 60
# Might timeout.
/usr/bin/time -v ./src/bitcoin-cli -datadir=$DATADIR loadtxoutset $SNAPSHOT_LOC || true
fg %1
```

After running this you should end up with a node that's synced to tip within 60-90 minutes depending on your internet connection and hardware.

### Benchmarking

I've compared reindexing to height 550,000 for this branch vs. its base ([a la previous PRs](https://github.com/bitcoin/bitcoin/pull/16801#issuecomment-551883348)), and beyond some outlier runs on both sides, this changeset seems to be comparable in performance to master.

```
host         tag                      time       time% maxmem  cpu%  dbcache
bench-ssd-2  bench/au.1               9:24:53    0.92  6717.99MB 338%  5000MB
bench-ssd-2  bench/au.1               9:23:14    0.92  6770.57MB 339%  5000MB
bench-ssd-2  bench/au.master.1        10:08:33   0.99  6654.39MB 315%  5000MB
bench-ssd-2  bench/au.master.1        10:13:04   1.00  6697.26MB 313%  5000MB

bench-ssd-3  bench/au.master.1        9:26:17    0.93  6743.38MB 337%  5000MB
bench-ssd-3  bench/au.master.1        9:24:03    0.93  6748.67MB 338%  5000MB
bench-ssd-3  bench/au.1               9:22:35    0.92  6713.50MB 340%  5000MB
bench-ssd-3  bench/au.1               10:08:36   1.00  6712.88MB 314%  5000MB

bench-ssd-4  bench/au.1               9:24:59    0.93  6744.43MB 338%  5000MB
bench-ssd-4  bench/au.1               10:06:09   1.00  6717.30MB 315%  5000MB
bench-ssd-4  bench/au.master.1        9:23:44    0.93  6695.61MB 339%  5000MB
bench-ssd-4  bench/au.master.1        9:23:56    0.93  6646.16MB 338%  5000MB

bench-ssd-5  bench/au.master.1        9:18:57    1.00  6715.23MB 341%  5000MB
bench-ssd-5  bench/au.master.1        9:20:35    1.00  6644.22MB 341%  5000MB
bench-ssd-5  bench/au.1               9:17:27    0.99  6702.63MB 342%  5000MB
bench-ssd-5  bench/au.1               9:21:17    1.00  6663.15MB 340%  5000MB
```

Note of course that as a sort of "meta"-optimization, these benchmarks are solely to ensure that we haven't introduced any performance regressions into normal operation. The really relevant benchmark here is "time from empty datadir to usable node," which on one of my more capable machines has gone from ~20hrs to ~1hr. 

### Next steps

If you want to help move this project forward, the following PRs are prerequisites and once they're merged I'll carve off the next changeset:

- #16945 
- #17487 

### Current work

After significantly revising the structure of the snapshot file in #16899, this branch has been substantially rebased to avoid the requirement of any on-disk snapshot metadata once the snapshot has been loaded; instead we look to filesystem state (the presence of a `$DATADIR/chainstate_*` folder) to signal that we are currently building a snapshot chainstate. Once this chainstate has been fully back validated, we move it over to the regular `chainstate/` dir on shutdown. We also used to need on-disk state to restore the "faked" `nChainTx` data for snapshot-based header chains -- we now cache that information in the snapshot's chainstate itself (https://github.com/bitcoin/bitcoin/pull/15606/commits/87093a94ba311c495a1982f4f47d0c4607789f8f).

I'm currently thinking about how to handle locking semantics (as you might be able to tell from the last commit, which needs to be squashed into the rest of the history at some point) which I can detail further if anyone's interested.